# Backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MediSync API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Register",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/recover-password": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Recover Password",
        "operationId": "recover_password_api_v1_auth_recover_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecoverPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Me",
        "operationId": "me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/api/v1/links": {
      "get": {
        "tags": [
          "links"
        ],
        "summary": "List Links",
        "operationId": "list_links_api_v1_links_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/VinculoDto"
                  },
                  "type": "array",
                  "title": "Response List Links Api V1 Links Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/links/request": {
      "post": {
        "tags": [
          "links"
        ],
        "summary": "Request Link",
        "operationId": "request_link_api_v1_links_request_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestLinkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VinculoDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/links/{id_vinculo}/respond": {
      "post": {
        "tags": [
          "links"
        ],
        "summary": "Respond Link",
        "operationId": "respond_link_api_v1_links__id_vinculo__respond_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_vinculo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id Vinculo"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RespondLinkRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VinculoDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/links/{id_vinculo}": {
      "delete": {
        "tags": [
          "links"
        ],
        "summary": "Revoke Link",
        "operationId": "revoke_link_api_v1_links__id_vinculo__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_vinculo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id Vinculo"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/devices/me": {
      "get": {
        "tags": [
          "devices"
        ],
        "summary": "Get My Device",
        "operationId": "get_my_device_api_v1_devices_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/DeviceDto"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get My Device Api V1 Devices Me Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "devices"
        ],
        "summary": "Unlink Device",
        "operationId": "unlink_device_api_v1_devices_me_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/devices/link": {
      "post": {
        "tags": [
          "devices"
        ],
        "summary": "Link Device",
        "operationId": "link_device_api_v1_devices_link_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkDeviceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/api/v1/alerts": {
      "get": {
        "tags": [
          "alerts"
        ],
        "summary": "Get Alerts",
        "operationId": "get_alerts_api_v1_alerts_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AlertaDto"
                  },
                  "type": "array",
                  "title": "Response Get Alerts Api V1 Alerts Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/alerts/{id_alerta}/read": {
      "post": {
        "tags": [
          "alerts"
        ],
        "summary": "Mark As Read",
        "operationId": "mark_as_read_api_v1_alerts__id_alerta__read_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_alerta",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id Alerta"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/alerts/read-all": {
      "post": {
        "tags": [
          "alerts"
        ],
        "summary": "Mark All As Read",
        "operationId": "mark_all_as_read_api_v1_alerts_read_all_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/alerts/{id_alerta}/acknowledge": {
      "post": {
        "tags": [
          "alerts"
        ],
        "summary": "Acknowledge",
        "operationId": "acknowledge_api_v1_alerts__id_alerta__acknowledge_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_alerta",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id Alerta"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertaDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AlertaDto": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "tipo": {
            "type": "string",
            "title": "Tipo"
          },
          "severidad": {
            "type": "string",
            "title": "Severidad"
          },
          "titulo": {
            "type": "string",
            "title": "Titulo"
          },
          "mensaje": {
            "type": "string",
            "title": "Mensaje"
          },
          "fecha_creacion": {
            "type": "string",
            "format": "date-time",
            "title": "Fecha Creacion"
          },
          "leida": {
            "type": "boolean",
            "title": "Leida",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "tipo",
          "severidad",
          "titulo",
          "mensaje",
          "fecha_creacion"
        ],
        "title": "AlertaDto"
      },
      "AuthResponse": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserDto"
          },
          "token": {
            "type": "string",
            "title": "Token"
          }
        },
        "type": "object",
        "required": [
          "user",
          "token"
        ],
        "title": "AuthResponse"
      },
...
      "DeviceDto": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "rssi": {
            "type": "integer",
            "title": "Rssi",
            "default": 0
          },
          "is_connected": {
            "type": "boolean",
            "title": "Is Connected",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "DeviceDto"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LinkDeviceRequest": {
        "properties": {
          "mac_address": {
            "type": "string",
            "title": "Mac Address"
          },
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "firmware_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Firmware Version"
          },
          "nivel_bateria": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nivel Bateria"
          }
        },
        "type": "object",
        "required": [
          "mac_address",
          "nombre"
        ],
        "title": "LinkDeviceRequest"
      },
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest"
      },
...
      "RecoverPasswordRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "email"
        ],
        "title": "RecoverPasswordRequest"
      },
      "RegisterRequest": {
        "properties": {
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "apellido": {
            "type": "string",
            "title": "Apellido"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "tipo_perfil": {
            "type": "string",
            "pattern": "^(paciente|cuidador|profesional_salud)$",
            "title": "Tipo Perfil"
          }
        },
        "type": "object",
        "required": [
          "nombre",
          "apellido",
          "email",
          "password",
          "tipo_perfil"
        ],
        "title": "RegisterRequest"
      },
      "RequestLinkRequest": {
        "properties": {
          "target_email": {
            "type": "string",
            "format": "email",
            "title": "Target Email"
          },
          "rol": {
            "type": "string",
            "pattern": "^(cuidador|profesional_salud)$",
            "title": "Rol"
          }
        },
        "type": "object",
        "required": [
          "target_email",
          "rol"
        ],
        "title": "RequestLinkRequest"
      },
      "RespondLinkRequest": {
        "properties": {
          "aceptar": {
            "type": "boolean",
            "title": "Aceptar"
          }
        },
        "type": "object",
        "required": [
          "aceptar"
        ],
        "title": "RespondLinkRequest"
      },
...
      "UserDto": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "apellido": {
            "type": "string",
            "title": "Apellido"
          },
          "tipo_perfil": {
            "type": "string",
            "title": "Tipo Perfil"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "nombre",
          "apellido",
          "tipo_perfil"
        ],
        "title": "UserDto"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VinculoDto": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "id_paciente": {
            "type": "integer",
            "title": "Id Paciente"
          },
          "id_usuario_vinculado": {
            "type": "integer",
            "title": "Id Usuario Vinculado"
          },
          "email_usuario_vinculado": {
            "type": "string",
            "title": "Email Usuario Vinculado"
          },
          "nombre_usuario_vinculado": {
            "type": "string",
            "title": "Nombre Usuario Vinculado"
          },
          "apellido_usuario_vinculado": {
            "type": "string",
            "title": "Apellido Usuario Vinculado"
          },
          "tipo_vinculo": {
            "type": "string",
            "title": "Tipo Vinculo"
          },
          "estado": {
            "type": "string",
            "title": "Estado"
          }
        },
        "type": "object",
        "required": [
          "id",
          "id_paciente",
          "id_usuario_vinculado",
          "email_usuario_vinculado",
          "nombre_usuario_vinculado",
          "apellido_usuario_vinculado",
          "tipo_vinculo",
          "estado"
        ],
        "title": "VinculoDto"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "api/v1/auth/login"
          }
        }
      }
    }
  }
}